Refactor TransactionTable
