+++ src/app/transactions/page.tsx
@@ -1,13 +1,10 @@
+import { Skeleton } from "@/components/Skeleton";
 import { Suspense } from "react";
 import { TransactionsTable } from "./TransactionsTable";
 
 export default function Transactions() {
   return (
-    <Suspense
-      fallback={
-        <div className="flex h-96 w-full animate-pulse rounded-lg bg-slate-200" />
-      }
-    >
+    <Suspense fallback={<Skeleton className="h-96 w-full" />}>
       <TransactionsTable />
     </Suspense>
   );
